+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-23
+- `17:51` -- **Updated** `apps/desktopProbe/src/index.ts`
 - `17:50` -- **Updated** `apps/desktopProbe/src/index.ts`
 - `17:50` -- **Added** `apps/desktopProbe/src/server/supabaseConfig.ts`
 - `17:40` -- **Updated** `apps/desktopProbe/src/index.ts`